dom/payments/test/test_payment-request-in-iframe.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1318988
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1318988</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript">

  "use strict";
  SimpleTest.waitForExplicitFinish();

  function testRequestInSameOrigin() {
    return new Promise((resolve, reject) => {
      var ifr = document.createElement('iframe');

      let listener = function(event) {
        is(event.data, "successful",
          "Expected 'successful', but got '" + event.data + "'");
        resolve();
      };

      window.addEventListener("message", listener);

      ifr.src = "simple_payment_request.html";
      document.body.appendChild(ifr);

      ifr.addEventListener('load', function() {
        window.removeEventListener("message", listener);
      });
    });
  }

  function testRequestInIFrame() {
    return new Promise((resolve, reject) => {
      var ifr = document.createElement('iframe');

      let listener = function(event) {
        is(event.data, "SecurityError",
           "Expected 'SecurityError', but got '" + event.data + "'");
        resolve();
      };

      window.addEventListener("message", listener);

      ifr.src = "https://test1.example.com:443/tests/dom/payments/test/simple_payment_request.html";
      document.body.appendChild(ifr);

      ifr.addEventListener('load', function() {
        window.removeEventListener("message", listener);
      });
    });
  }

  function testRequestInIFrameWithAttribute() {
    return new Promise((resolve, reject) => {
      var ifrr = document.createElement('iframe');

      window.addEventListener("message", function(event) {
        is(event.data, "successful",
          "Expected 'successful', but got '" + event.data + "'");
        resolve();
      });

      ifrr.setAttribute('allowpaymentrequest', '');
      ifrr.src = "https://test1.example.com:443/tests/dom/payments/test/simple_payment_request.html";
      document.body.appendChild(ifrr);
    });
  }

  function runTests() {
    testRequestInSameOrigin()
    .then(testRequestInIFrame)
    .then(testRequestInIFrameWithAttribute)
    .then(SimpleTest.finish)
    .catch( e => {
      ok(false, "Unexpected error: " + e.name);
      SimpleTest.finish();
    });
  }

  window.addEventListener('load', function() {
    SpecialPowers.pushPrefEnv({
      'set': [
        ['dom.payments.request.enabled', true],
      ]
    }, runTests);
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1318988">Mozilla Bug 1318988</a>
</body>
</html>
